update doc & examples

+++ examples/06_manipulating_images/plot_affine_transformation.py
@@ -77,7 +77,8 @@ source_affine[:2, :2] = rotation_matrix * 2.0  # 2.0mm voxel size
 
 # We need to turn this data into a nibabel image
 import nibabel
-img = nibabel.Nifti1Image(image[:, :, np.newaxis], affine=source_affine)
+img = nibabel.Nifti1Image(image[:, :, np.newaxis].astype("int32"),
+                          affine=source_affine)
 
 #############################################################################
 # Now resample the image
